Create a voter TaskRemoveVoter, add annotation authorisation on delet…

+++ src/Controller/TaskController.php
@@ -7,6 +7,7 @@ use App\Form\TaskType;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 class TaskController extends AbstractController
@@ -78,12 +79,16 @@ class TaskController extends AbstractController
 
     /**
      * @Route("/tasks/{id}/delete", name="task_delete")
+     * @IsGranted("REMOVE", subject="task")
      */
     public function deleteTaskAction(Task $task, EntityManagerInterface $em)
     {
+        if (!$this->isGranted('REMOVE', $task)) {
+            throw $this->createAccessDeniedException();
+        }
+
         $em->remove($task);
         $em->flush();
-
         $this->addFlash('success', 'La tâche a bien été supprimée.');
 
         return $this->redirectToRoute('task_list');
